requirements.txt
fastapi~=0.115.13
uvicorn~=0.34.3
requests~=2.32.4
PyYAML~=6.0.2
pydantic~=2.11.7
pycryptodome~=3.23.0
PyJWT~=2.10.1